requirements.txt
streamlit
llama-index
llama-index-llms-google-genai
google-genai
llama-index-embeddings-openai
llama-index-vector-stores-pinecone
pinecone-client

torch
sentence-transformers
ragas
pandas
python-dotenv
langchain-openai
langchain-community
unstructured[pdf]
trulens-core
trulens-apps-llamaindex
trulens-providers-openai

